requirements.txt
requests
beautifulsoup4
tweepy
selenium
pillow